Split use mint
