+++ hooks/useMintCandyMachine.tsx
@@ -1,80 +1,24 @@
 import { useAnchorWallet, useWallet } from '@solana/wallet-adapter-react'
-import { AnchorProvider, Program, BN } from '@project-serum/anchor'
+import { AnchorProvider } from '@project-serum/anchor'
 import { useRPC } from 'hooks'
-import { PublicKey, LAMPORTS_PER_SOL, Transaction, Connection } from '@solana/web3.js'
+import { Transaction, Connection } from '@solana/web3.js'
 import { awaitTransactionSignatureConfirmation } from 'lib/candy-machine/upload/transactions'
 import { useState } from 'react'
-import { CANDY_MACHINE_PROGRAM_V2_ID } from 'lib/candy-machine/mint/constants'
 import { DEFAULT_TIMEOUT } from 'lib/constants'
-import { SetupState, CandyMachineAccount } from 'lib/candy-machine/interfaces'
+import { SetupState } from 'lib/candy-machine/interfaces'
 import { createAccountsForMint, mintOneNft } from 'lib/candy-machine/mint/mint'
+import useRefreshCandyMachine from './useRefreshCandyMachine'
 
 const useMintCandyMachine = (account: string) => {
     const anchorWallet = useAnchorWallet()
     const { connection, network } = useRPC()
     const [isUserMinting, setIsUserMinting] = useState(false)
-    const [itemsRemaining, setItemsRemaining] = useState(0)
-    const [itemsAvailable, setItemsAvailable] = useState(0)
-    const [nftPrice, setNftPrice] = useState(0)
-    const [isActive, setIsActive] = useState(false)
-    const [isCaptcha, setIsCaptcha] = useState(false)
     const [setupTxn, setSetupTxn] = useState<SetupState>()
     const [mintMessage, setMintMessage] = useState({ error: false, message: '' })
-    const [candyMachine, setCandyMachine] = useState<CandyMachineAccount>()
+    const { itemsRemaining, refreshCandyMachineState, candyMachine, setItemsRemaining, setIsActive, isCaptcha } =
+        useRefreshCandyMachine(account as string)
 
     const wallet = useWallet()
-
-    async function refreshCandyMachineState() {
-        try {
-            if (!anchorWallet || !wallet.publicKey || !connection || !network) return
-            const provider = new AnchorProvider(new Connection(connection.rpcEndpoint), anchorWallet, {
-                preflightCommitment: 'recent',
-            })
-
-            const idl = await Program.fetchIdl(CANDY_MACHINE_PROGRAM_V2_ID, provider)
-
-            const program = new Program(idl!, CANDY_MACHINE_PROGRAM_V2_ID, provider)
-            const state: any = await program.account.candyMachine.fetch(new PublicKey(account))
-            const itemsAvailable = state.data.itemsAvailable.toNumber()
-            const itemsRedeemed = state.itemsRedeemed.toNumber()
-            const itemsRemaining = itemsAvailable - itemsRedeemed
-            const captcha = !!state.data.gatekeeper
-            let nftPrice = new BN(state.data.price).toNumber() / LAMPORTS_PER_SOL
-            let active = new BN(state.goLiveDate).toNumber() < new Date().getTime() / 1000
-
-            setItemsRemaining(itemsRemaining)
-            setItemsAvailable(itemsAvailable)
-            setNftPrice(nftPrice)
-            setIsActive(active)
-            setIsCaptcha(captcha)
-
-            setCandyMachine({
-                id: new PublicKey(account),
-                program,
-                state: {
-                    authority: state.authority,
-                    itemsAvailable,
-                    itemsRedeemed,
-                    itemsRemaining,
-                    isSoldOut: itemsRemaining === 0,
-                    isActive: false,
-                    isPresale: false,
-                    isWhitelistOnly: false,
-                    goLiveDate: state.data.goLiveDate,
-                    treasury: state.wallet,
-                    tokenMint: state.tokenMint,
-                    gatekeeper: state.data.gatekeeper,
-                    endSettings: state.data.endSettings,
-                    whitelistMintSettings: state.data.whitelistMintSettings,
-                    hiddenSettings: state.data.hiddenSettings,
-                    price: state.data.price,
-                    retainAuthority: state.data.retainAuthority,
-                },
-            })
-        } catch (err) {
-            console.error(err)
-        }
-    }
 
     const mintAccount = async (beforeTransactions: Transaction[] = [], afterTransactions: Transaction[] = []) => {
         try {
@@ -157,14 +101,9 @@ const useMintCandyMachine = (account: string) => {
     return {
         mintAccount,
         isUserMinting,
-        itemsRemaining,
-        itemsAvailable,
-        nftPrice,
-        isActive,
-        refreshCandyMachineState,
         mintMessage,
+        itemsRemaining,
         isCaptcha,
-        setIsCaptcha,
     }
 }
